requirements.txt
# torch
# transformers==4.48.0
# Pillow
# gradio==5.34.2
# huggingface_hub==0.25.2
# sentencepiece
# accelerate
# protobuf
# opencv


# gradio==5.34.2
# google-generativeai==0.4.1
# Pillow
# python-dotenv

# gradio==4.25.0
# google-generativeai
# python-dotenv
# Pillow
# sqlite-utils

# langchain>= 0.1.15
# langgraph
# langchain-google-genai
# google-generativeai
# gradio
# python-dotenv

langchain>=0.1.16
langgraph>=0.0.33
langchain-google-genai>=0.0.7
google-generativeai>=0.4.1
gradio>=4.19.2
python-dotenv>=1.0.1


